Add variables clean_arguments and clean_commands
It looks better in \echo if I prevent mentions by using
message.clean_content, instead of replacing every instance of '@' with
'@' followed by the zero-width space character.

To do that, I have to add "clean content" versions of
message.clean_content for the variables arguments and command in
on_message().