Fix exception when running \countallpins if no access to a channel
On Sunday, the 8th of January, 2017, at 11:04 CET, a member by the username of
VVVVVV#6391 (at the time of writing having the nickname "no sixteen" on
the Try Not To Get Banned server) decided to run the command \countallpins
in the channel #only-use-the-bot. One would expect the output of such
command to be a list of all channels on the server the command is run on,
accompanied by the length of the set of messages for which the properties
hold that the message was sent in that channel, and the message is also an
element of the set of messages for which holds that for every message in
that set, that message is a pinned message, and also 50 minus said length
(which is the same number as the amount of messages that can still be
pinned in that channel, due to restrictions of Discord that only 50
messages may be pinned in a given channel). However, that is not what
happened, given that the bot did not reply to that command at all.
However, an error message was output to the console, only accessible by
Info Teddy, the person who is running the bot on his personal computer,
that indicated that a permission error had occurred (specifically
discord.errors.Forbidden: FORBIDDEN (status code: 403): Missing Access).

We decided that it was necessary to debug this problem, and as such, we
read the error message. I (Dav999) was puzzled as to why the bot did not
have access to perform the task that it was requested to do via the
aforementioned command, as expressed via the message I sent in
the #bracketed-backslash channel on the Open Level Project server on
Sunday, the 8th of January, 2017, 21:12 CET, with the contents "How is
that possible". It was clear that the bot was not restricted in sending
the response message, even though the error message would appear to
indicate that the cause of the problem was such. At a later point in time,
it became evident that there existed a channel for which held that that
channel was a channel on the Try Not To Get Banned server and the bot [\]
did not have access to that channel. Needless to say, the bot was unable
to retrieve the list of pinned messages in such a channel, causing the
execution of the command to fail. The solution would be to insert a
try-catch block, catching any discord.errors.Forbidden exception, and
when caught, causing the display of not the aforementioned length of the
set of pinned messages for that channel and of 50 minus said length, but
of a note with the intention to convey the message to whichever reader of
said message, that retrieving said length and 50 minus said length was
unsuccessful. The intention of this commit is to make this change. Thus,
if applied, this commit will fix the exception when running
\countallpins if the bot does not have access to any of the channels.

You may wonder what I want to accomplish by writing such a long commit
message. Well, Info Teddy decided to like having a consistent strict
commit message style, and mocked my personal style for being hard to read,
calling it "atrocious". I decided to have some fun by mocking his new
style of always commenting below the commit message subject line. It was
fun to write this, even though it wasted some of my time, so I would
appreciate if someone would read all of this.