Rewrite commands system
Rewrote the commands system to use a new function, shadow(), and then
made all of the commands use it as a function decorator, like so:

  @shadow()
  async def help(client, message, **kwargs):

This is more readable than a string of if-elif-...-elif-else, for two
reasons:

  - One less indentation
  - Commands as individual functions make their auth and aliases
    clearer, and as a bonus, if a bug happens to occur in the command,
    the traceback will nicely list the command instead of having you
    guess which command the error was in

All commands previously utilizing the command and arguments variables
will now use kwags['command'] and kwargs['arguments'] respectively.

Command names cannot be the namespace of something else without causing
errors, although you could set an alias to them as a workaround. A
likely solution for this would be to put all of the commands in a
separate file, and then import the file.

Function decorators do not appear to know the existence of functions
that were exec(compile(open()))'d, maybe because that is done at the end
of the file. So, I had to copy some of the auth functions from
functions.py into main.py for the auth functions to work.

Another thing changed in this commit is that it makes some parts of the
code style more consistent.