Make bot not cache DM attachments and image embeds
Makes the bot not cache attachments and image link embeds from direct
messages. This is because attachments and image link embeds are cached
for if the message sender happens to delete the attachment later, but if
the bot doesn’t log message deletions in direct messages, then there is no
need to cache them.