Make role cache warning not show up for members without roles
This makes it so a member is no longer saved to the role cache with 0
roles if they're in the join channel while the bot restarts.

Also, the bot was supposed to delete all messages in the join channel, not
just the commands. I should probably also check for mod/admin powers, to
make them able to post without changing the config.