Make reply() send terminal output if failed
Made reply() catch if sending the message failed and returned either a
discord.errors.HTTPException or discord.errors.Forbidden, and then
outputs the rejected message to the terminal, and the server and channel
it was attempted to be sent to.

This is done so the host can examine exactly what went wrong with
sending the message.