Add var messages_deleted_by_bot and usage of it
Added a global variable named messages_deleted_by_bot, which is a list
of messages deleted by the bot, and made every client.delete_message()
the bot calls also append the message deleted to the global variable.

Now, inside the code block in on_message_delete() for resending recently
deleted messages, the bot now checks if the message is one of the
messages the bot deleted. If it finds that that's true, it deletes it,
and removes the message from the variable.

This is done because it's annoying when the bot resends a message that
it itself deleted for a (presumably) good reason.