Change text in deleted message notification when content is not resent
It said "UNDELETED MESSAGE" and "This message was resent as it was deleted
too recently", which is a little weird if we're not actually resending the
message contents. So now that has a title "Message was deleted" and
description "This notification was sent because a message by this user was
deleted too recently." instead.