Revert "Add some type checking when loading the config"
This reverts commit bf4bfb660a2e19156d0e1837fa46adce37f96a82.

To hell with this, stuff like this shouldn't happen unless someone edits
the code who really doesn't know what they're doing, it adds code that has
become ugly because it has to fit in 100 columns, it adds complexity and
unreadability for something that can probably be detected easily enough.
Unless Info thinks it's a good idea to add it anyway.