config.py: Add notify_invalidcmd config option
Added a config option named notify_invalidcmd. This sets whether the bot
should message on an invalid command, instead of failing silently. Note
that this commit only adds the config option; the code to actually fail
silent has not been implemented yet.

If [\] is to go public, I expect that people would likely not want the
bot to send a message on an invalid command, as most people expect the
bot to not do that, as most bots do not do that and meew0's "Best
practices for Discord bots" document recommends to not do that, too.