main.py: Add support for disallowed func chars in command func names
For example, putting "__002F" in a command function's name is equivalent
to simply adding a forward slash in its name.

This was done in order to be able to put in the rest of the commands
which had names that could not be the name of a function, into the rest
of commands.py, for consistency with every command.