functions.py: Add is_host() auth check function
There are going to be commands in the future that should be bot
host-only, that are not simply eval and its friends. This is also for
consistency with the rest of the auth functions.

Name is is_host() instead of is_owner() to ensure no confusion occurs
over if the function is checking for server ownership status or bot
ownership status.