commands.py: Add \sudo
Sometimes an operator will have to do a command, even if he does not
have permission to do so normally. \sudo is functionally the same as the
sudo command on *nix systems, except you don't have to put your password
in, and sometimes it won't execute because, even still, the bot doesn't
have perms.