main.py: Add servonly parameter to shadow()
The servonly parameter specifies whether or not the command can only be
run inside a server, i.e. not allowed to be run in a private message.

Reusing "if privatemessage: e = emb.error(); await reply(); return"
so much in commands means it's time to make it so that you don't have to
keep copying and pasting that stuff.