main.py: Fix voice text chan bug with user muting
If a user got muted or deafened somehow while in a voice voice channel
linked with a voice text channel, that user's Read Messages permission
would be changed for that channel. This was caused by not checking
whether or not the voice channels were the same in the client event
on_voice_state_update().