commands.py: Make expirylist() usable in DMs
expirylist() now takes an argument for identifying a server with
utils.match_input() when run inside direct messages. It only returns
servers the command caller is actually on, so don't worry about leaking
data or anything. However, it can be overridden with the "sudo" command
(i.e. only bot operators can use it) to forcefully show servers the
command caller isn't on.