utils.py: Rewrite mdspecialchars()
The approach of mdspecialchars() has been changed so that it has an
iterable of characters that are not special characters, and now it just
escapes every character that's not in that iterable. Now it escapes
emojis without having to refer to a list of emojis that will always be
constantly changing and being in need of updates.

This was done because having emojis that wouldn't be escaped and were
being displayed as full pictures in log messages was getting annoying.