utils.py: Fix mdspecialchars() backslash bug
If there were multiple special characters in the string that needed to
be escaped, mdspecialchars() would put a backslash every time. So, for
example:

Input: '_____________'

Output: '\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_\\\\\\\\\\\\\_'

This was because it was looping over its own backslashes that it itself
put in. By making it so that it uses a new variable instead mutating the
input string, this bug is fixed.