config.py: Add option tntgb_servers
This is to make the bot more universal and not only allowed on specific
servers. Servers which have TNTGB enabled will be able to use TNTGB
commands, among other things.

This is an array of server IDs that have TNTGB enabled, and not a
detachable boolean, because it's easier to see which servers have TNTGB
enabled and which don't with an array. Also, it is a hidden config
option, as it could reveal servers which people are not on, and thus
leaks data.