Add `\archive`
This command can be used to get a simple chat log of the latest X messages
in a given channel. It accepts a channel representation, and optionally a
limit of messages. There's a maximum limit set in a non-detachable config
option, in case server owners need to be prevented from bringing the bot
down by requesting hundreds of thousands of messages.

We may want to make this command available to non-admins in the future,
but for now it is admin-only, particularly so that we can think about read
and history permissions later.

This does not log attachments, rich embeds or reactions (yet?), or
anything fancy with IDs. It just gives a human-readable file with for each
message a timestamp (in configured date format), a username with
discriminator, and the message content.

Syntax checked