main.py: Make reply() except handling use better log interpolation
Make reply() use, for logging functions, % formatting interpolation, but
passing the arguments of the % formatting interpolation as normal
arguments to the logging function.

pylint was complaining about it, and it saves performance by handing off
the text interpolation until the last second, so if the logging level
doesn’t make the logging function actually log, resources aren’t being
wasted formatting a string for nothing.