commands.py: Refactor b()'s valid member checking
The function was originally written with a try-except block that was
catching AttributeErrors and TypeErrors from a giant range of code,
which increased the chances that the bot would say the caller had
targeted the wrong member, when in fact, that was not the case.

This makes it so b() has better checking for an invalid member, as in,
it actually checks that the member is None instead of guessing that
AttributeErrors and TypeErrors would always signify an invalid member.