Refactor the bot
This moves the bot structure around even more. Also, some other things
like code style issues have been fixed.

There is now a bot.py, which contains all the client-related functions.
This is a more logical grouping, and it makes reloading parts of the
bot's code easier. Also, it gets rid of importing __main__, which it
easier to check code with a linter, without having to modify code or
create new files to make the linter understand the structure of the
code.