Change expiry types for custom role commands (`input_strict` and `input`)
The existing `input` type was renamed to `input_strict`, because it always
requires an expiry time to be given.

The new `input` type doesn't require the time to be given, if not given or
invalid, it will automatically not be set to expire. So if you use a
command like `\ban Aaa Bbb`, and the expiry type is set to `input`, then
"Aaa Bbb" will be seen as the name since "Aaa" is not a valid expiry time,
and it will be a non-expiring ban (or at least until you lift it). If you
use `\ban 2h Aaa Bbb`, "2h" is a valid expiry time, so the name will also
be "Aaa Bbb".

Existing custom role commands will *not* be mass-changed to `input_strict`

Also added some comments to the run function.