Fix \b_id bug 'built-in not in role cache'
Fix a \b_id bug where the command would error with 'Member <built-in
function input> cannot be found in the role cache. Please note you have
to enter an ID, not any form of name!'

The direct cause was an incomplete refactoring of the function to not
redefine built-ins, but the actual cause was redefining built-ins, which
causes bugs linters cannot account for.