Port the bot to discord.py rewrite
discord.py rewrite has a lot of useful new features, and many other bots
are being ported to it.

Changes to the bot include, but are not limited to:

 - Indefinite typing when computing a task that could potentially take
   a while, instead of one typing event
 - Referring to servers as guilds internally
 - Avatar change logs explicitly logging default avatar URLs
 - Discord IDs (snowflakes) being integers instead of strings, except
   when stored as key indices in JSON
 - isinstance checks of channels instead of checking their type
   attribute or is_private attribute

logids.txt has been updated to refer to servers as guilds, and readme.md
has been updated to refer to the new version of the discord.py
dependency.