Rewrite the hangman system
Hangman desperately needed to be changed, since it added half a dozen
global variables and was only supported in one, hardcoded, channel.

Most importantly, there's now a hangman.py containing a HangmanGame class,
which represents one instance of hangman. That way, you can have multiple
games running at the same time in different channels. hangman.games is a
dictionary with channel IDs as keys and HangmanGame objects as values.

The idea is that \hangman should be used in the channel itself, and then
the bot sends a DM to ask for the word. This is not yet implemented
however, \hangman should, for now, be passed the word in the same channel
the game is to be played in.