Implement a reload command
Move the client instance to a new wrapper.py, that won't be reloaded at
all. If the client instance is reloaded, it will have to reconnect to
Discord, which defeats the purpose of reloading. Edit every direct
reference to the client instance to point to its new location in
wrapper.py.

Add a reload_bot() to main.py that is the function to call when
reloading the bot. Add a recursive_reload(), which reload_bot() calls.

recursive_reload() goes through each module member of an internal bot
module and reloads it if it hasn't already been reloaded. It
specifically does not reload the modules of 'bot', 'wrapper', and
'main'.

Add a command to call reload_bot().