Fix 3 bugs relating to joinchannel and role gives
The first bug was the bot refusing to respond to 'join' commands in a
guild's join channel, because of duplicated code.

The second bug was using the 'send' method of a member instead of
'add_roles'.

The third bug was a block of code that was indented one level too much.