Fix 3 bugs for b_id()
The first bug was the bot not properly rolecache-rolebanning people,
which was caused by not properly converting an ID input to int.

The second bug was the expiry entry not being properly set and breaking
the saved JSON file caused by not completely renaming all instances of
redefining the built-in name 'input'.

The third bug was the bot always banning a member for 5 days, which was
caused by not taking into account the TNTGB config options.