Make msg_start not a global
Global variables are evil or something.

This abstracts calculation of msg_start to a function, so everything
else can get a msg_start when it needs to. This also means that the bot
can no longer reply with the wrong input display, because there won't be
any race conditions or anything.