Fix bug in match_input() matching discriminators
The bot didn't check the length of the input to see if the caller was
trying to match a discriminator, and so seeing that it was composed
entirely of digits, assumed it was a snowflake ID. This meant it would
break when the request moves down gets passed through over parts of the
functions. Add a check to see if the length of the string matches the
length of the discriminator.