Add 0-len content handling for message edits
The older and newer content of message edits are placed in rich embed
fields, which are required to have values. This meant if a user or bot
managed to send a message with no content that had either a rich embed
or attachment, that user or bot could cause the app to get an exception
by editing in content or editing out content.