Clarify when command "faults" are raised
When a "fault did occur" in a command, a line with a red background is
now printed before the error is raised. This makes it clear when
someone gets a fault message when trying to run a command in a valid
way, and sets the error apart when there are other unimportant lines.