Add \move command
It allows moderators to redirect people to a different channel to
continue the current conversation in, and includes links to jump
between the parts.

Only problem is [\] is failing to use custom emotes for some reason,
and yes, I am using <:name:id> format, yes, the bot is in the servers
that have emotes I'm trying to use, and the same thing does work with
] correctly. I'll just use 🚦 and ⤵ instead of red and green traffic
lights, I guess.