Make command prefixes not be loaded by reference
So this is pretty funny if you think about it. Basically, every time
a message is sent in a server that has a detached list of prefixes,
the master prefixes list is extended with the local prefixes so both
can be used, but it's done by reference. So local prefixes will
suddenly become useable in every server, and the list of global
prefixes will grow bigger over time as messages are sent in servers
with custom prefixes that keep getting added to the global list. Yes,
this is also saved to the config file.