Adapt raw events in advance of future starboard feature
Plans for an (improved) starboard feature exist, and I'm planning to
use the raw events for more robustness (uncached messages being
ignored is always a gotcha).

This commit makes sure the logging functionalities being disabled
doesn't actually return from the entire event, now the conditional
determines whether the logging code should be run or not without
returning. This is so the event can be easily used for more things.