Add start of starboard implementation
A start was made with the code to run when a message is starred/
unstarred/nostarred/unnostarred. This uses raw events, as to be
reliable, but still checks messages in the cache instead of fetching
them from Discord when possible.

A lot of config options were added, and a `rti` config type: a number
of seconds, which can be input by the user as a relative time like
2d or 1h30m.