Implement rest of "starworthiness" checking code
All the code that decides whether something should be on the starboard
and with how many stars, based on configuration and how people star
their own messages and so on, is now finished.

To do: implement actual starboard handling code, like, make the
announcements, edit them, remove them, have the correct counts and
color, etc.

I'll probably detail all the design decisions in a wiki article.