Fix potential race condition regarding starboard-banned users
Scenario: Assume the threshold is 5 stars. Someone would add the 4th
star to a message, very shortly afterwards a starboard-banned user
adds a star, which should not be counted. The event for the 4th star
is being handled, but counts the 5th star by the starboard-banned user
before it can be removed by the event handling that star. Thus, the
message would get starboarded.

This would be caused by the adding-up of reactions not taking
starboard-banned users into account, as to not remove older stars that
users have made while not yet being starboard-banned.

The solution/mitigation is to create a temporary record in the event
for the starboard-banned user's star/nostar, saying a starboard-banned
user just added a star/nostar to a specific message and that it should
not be counted if it ends up in a list of starrers/nostarrers for that
message in another simultaneous star event handling.