Fix mishap when trying to set an external emote as star/nostar
It was supposed to be blocked, and the documentation implied that
Reaction.emoji could only be Emoji or str, implying it couldn't be
a PartialEmoji, which it actually can be.

I think the same applies when a message actually contains any
PartialEmoji, so fixed it in the star checking code too.