Add logging of server boosts and unboosts
This depends on the Nitro Booster role. The config option
`nitrobooster` is used to store the booster role ID for a server, which
will be set automatically if a managed role is created called "Nitro
Booster". The logids `member_boost` and `member_unboost` control
whether or not adding or removing the booster role is logged,
completely independent from the existing `member_roleadd` and
`member_roleremove`. (So you could log when someone boosts or unboosts
the server without logging all role adds/removes, or you might want to
log role adds/removes but don't want to log boosts/unboosts.)