Improve robustness against denied access to send messages in starboard
When making the starboard, people should deny everyone the right to
send messages in the channel, but may easily forget to allow the bot.
Now, when the bot can't send a message to the starboard, it will react
with a 🔑 emote on the original message, indicating a permission error.

There was also a bug where this permission error would leave a message
pretty much stuck in the state "already being starboarded", thus
preventing it from being starboarded after the permission error has
been solved. This is now fixed.