Add new command \channelperms
This command can be used by anyone with the permission to manage roles,
and gives a good overview of access permissions, either
1) all permission overrides for all channels or
2) the evaluated permissions for a specific member in all channels.
Note that only the permission to read messages and send messages is
shown. Also, for the best view that was still somewhat reasonable to
make within Discord's possibilities, the overview is uploaded as an
HTML document that can easily be viewed in your browser.