Add access level properties to all config options
This will control which options can be seen and edited by local admins.
The sysop options will be hidden, some options will be visible but
read-only because another command already controls them and already has
checks built-in. Right now only the starboard and TNTGB options will be
read-only (except for the starboard bans because that could be used to
ban someone who has left the server). Later, these options could be
changeable via \config. Especially if the new GUI possibilities that
were showcased by Discord will become a thing.

(I'm really looking forward to that by the way, but I'm also left to
wonder if this isn't going to be only possible for verified bots (where
verification is, of course, gotcha, only available if your bot is in at
least 75 servers). I see official messages vaguely hinting in both the
direction that verification is going to be necessary to "release
features that give developers more control over Discord" and that we
should "rest assured that we know how valuable these features will be
to your work regardless of your bots' scale." We'll see, I guess)