Make startup sequence more robust, add "Starting..." and error status
Earlier today Discord went down (or rather, half the internet
centralized on Cloudflare) and when the bot could go back online, it
tried to send a connection message to the operating server and failed
because that server still had an outage. As a result the rest of the
on_ready function wasn't run, including role cache initialization and
role expiry initialization... A similar thing happened long ago when
the bot started up after a long uptime and a server turned out to have
been deleted while it was looking through servers to check for the role
cache or whatever.

Now the on_ready function is split into different `try` blocks, so
failure loading one part doesn't affect the other. For example, if the
connection message cannot be sent to the operating server, it will
still try to separately initialize role cache and expiry and rules.
It can indicate failure of specific parts via an error code in the
bot's status. Later on, we can close off certain functionality if a
part failed to load (don't add new role expiries to an empty list of
expiries and give error messages when trying to use that system, for
example). Also, during startup, the status now shows "Starting..."

Still with "Playing" before it because Discord. Seriously, bots were
precisely the ones in need of custom statuses (without "Playing"), and
we still can't use it? What's so hard about just adding a new activity
with no compulsory text ("Playing", "Watching") then, as an
alternative? Does it need to be over-engineered BUT we have nothing
during the years before it's there? (remember search)