Attempt to improve \restart command
Looks like it was calling a new python2 process (giving a syntax error)
and also didn't stop itself. I now made \restart exit with code 8, and
I also changed the exit code for \kill from 42 to 4, and adapted the
shellscript for it to treat those codes differently. Why 4 and 8? Well,
42 was a bit too "funny", and 4 sounds like "death" in certain Asian
languages (which is why they skip 4th floors). 8 is like an infinite
loop. Could've chosen any numbers, but it'd help making them mnemonics.