Remove "GPLv3 or later" and make it GPLv3-only
I don't like that the FSF recommends you use GPLv3 or later. Their
justification is that this is done to automatically upgrade your license
to future versions, to protect against possibly unforeseen security
holes in the current version of the license. The fact that this
automatic upgrade adds the **ENTIRETY OF THE FREE SOFTWARE FOUNDATION**
to the attack surface of the license doesn't seem to have occurred to
them.