Account for 0 text channels in findc
After a (pretty old by now) Discord update, you're now able to have 0
text channels in a server. Although if this is the case, then I'm not
sure how the findc message got sent and received by us in the first
place, but I'll put this down just in case it somehow happens.