De-duplicate use of utils.paginate() with embeds
The standard pattern is this: we have a newline-separated list of
things. However, it's possible for this to go over the embed field value
limit of 1024 chars. So if it is, call utils.paginate(), and insert a
new field for each page (with a title of U+200B). Else just add the
embed as normal.

In order to prevent copy-pasting code, I've refactored this into
utils.paginate_field().