Rework role change logs for singular/plural add/remove and mixed cases
So this is what the bot will output now for each case:

If it's a singular role add/remove, it will put the role name in the
embed description and role ID in the footer summary. There will be no
fields in the embed.

If it's a plural role add/remove, it will put the role names in a list
in the embed description (and add more fields to paginate, if
necessary). Again, there will be no fields in the embed (if it doesn't
need to paginate).

If it's mixed (i.e. there are both roles added and removed
simultaneously), then it'll have fields that list both the role(s) added
and removed (both of which will be paginated appropriately, of course).

Additionally, 🔀 (twisted rightwards arrows) is the emoji for mixed, 📥
(inbox tray) is the emoji for adding roles, and 📤 (outbox tray) is the
emoji for removing roles. Of course, they will all be preceded by the 🔑
(key) emoji.