Abstract "bot connected" msg to function and prettify on_ready output
This is just me committing a stash from probably years ago, where I
prettified the output that gets sent to the terminal, as well as
abstracted the "bot connected" message to a function. Even though it's
not duplicated, it's still probably good to put it in its own function.
Also, I can replace that raw emoji with \N{EMOJI} syntax instead, and
add a timestamp to the embed.