Add manual jump links to utils
It'd be nice to be able to just use the 'jump_url' property of
discord.Message objects, but unfortunately, it's a property, so we can't
make our own object and pass it in to that property, as Python will give
an error if we try to do so.

This adds utils.manual_jump_link(), utils.get_manual_jump_link(), and
utils.embed_manual_jump_link().