De-duplicate converting channel types to names
This has been abstracted to utils.get_channel_type(). Furthermore, the
output of this function will be mixed-case, in case we want the type of
the channel but aren't using it in the titles of log messages.

I've also decided to use the built-in enums instead of directly relying
on the types themselves, as that's what the enums are meant to be used
for.