Add core of reaction roles system (untested)
Reaction roles is a system where you're able to assign yourself a role
in a guild by simply reacting to a message. Usually this is in a channel
the guild administrators have set up where you can't send messages and
you can't add new reactions, but you can react to existing ones. And if
you unreact to a role, your role gets removed accordingly.

Currently, all this adds is the core part of the system, which is simply
just making the database, and looking up it database and giving users a
role if they react with a specific emoji on a specific message. Other
features still need to be added, such as:
 * An interface to let guild administrators actually set up the reaction
   role system
 * A interface to make [\] automatically send a message listing each
   emoji and its assigned role, if you don't want to make the message
   yourself
 * If you do make the message yourself, an interface to assign a set of
   roles to that message

This commit is basically untested. I've only tested the syntax and the
database creation, which is the bare minimum. I've not actually tested
that the role adding or removing works. But this is the untested branch,
so I'm committing and pushing right now instead of leaving my directory
dirty when I go sleep tonight. So screw it.