Don't delete all roles from every message
Just realized that I changed the function so that it would remove all
role IDs from every message... maybe a guild admin wants a role across
multiple messages. It'd be weird (it'd allow for the user to not have a
reaction to a message but still have a role, breaking down an already
fragile correlation) but it shouldn't be suddenly deleted.

Anyway, it's fixed now. Furthermore, you need to spell out the keyword
arguments to the function.