Indicate NSFW channels
This is done by prepending "NSFW" if the channel is NSFW, and adding the
underage emoji (🔞) after all the other emojis.

Apparently not every channel has the is_nsfw() attribute, but apparently
categories can be NSFW too, but this isn't exposed in the UI... but why
can't voice channels be NSFW as well? Whatever, just Discord things.