Modernize channel creation/deletion logs
Before, they were primitive and placeholders. But now they're all fancy
like the rest of the logs are.

I added utils.get_kbps() because it would be a bit too verbose to do the
logic inline in the voice channel creation log.

Channels could have specific permissions when they get created. I
haven't added code to show this in the creation log, but I'll do that
later as it's a bit more complicated than the other attributes of a
channel.