Use mutee mention in \vy and \vn
Before allowed_mentions was added, we used 'NAME#DISCRIM (ID)' to
prevent mentioning the mutee. However, since we can disable the mention,
it's better to just use the mention as it's less complicated than
constructing a string ourselves.

The only reason to use 'NAME#DISCRIM (ID)' would be in cases where we
can't rely on a mention or want to be absolutely sure the user account
always shows up properly instead of relying on user client caching, and
this is not one of those cases.