Put ID summary in footer and remove "(temp)" for reaction clear logs
Apparently the message ID was being stored in a field called "Message ID
(temp)" instead of using the proper ID summary. So I properly put the
message ID in the footer, along with the channel ID.

In the future we ought to have timestamps on every log message and take
all ID summaries out of the footer, but this will be fine for now.