Add \add_role_all and \remove_role_all
Just a simple thing to automate adding and removing a role from everyone
in the server, because yet again Discord doesn't have that feature.

commands.py now imports dispatch.py, the string 'specify_role' has been
added, and the commands are in the help list as well. utils.obj_info()
has a new parameter signifying whether to disable the Markdown output or
not (used in specifying the reason in the audit log).