Add nologusers config option
This config option can be used to prevent edits and deletes from being
logged for messages by certain users.

This may be useful if, for example, you have a bot that regularly
replies to commands with a "Processing..." message and then edits the
result. In our case, a Minecraft bot keeps sending and editing messages
showing inventories and items.