Temporarily remove support for gateway intents
We're running into some issues with 1.5.0, particularly messages not
being hashable anymore, thus not being directly comparable anymore.
Apparently they never were supposed to be, but we assumed it would be
good practice to compare Message objects directly than to compare IDs
or apparently jump_urls as people are suggesting here:
https://github.com/Rapptz/discord.py/issues/5866

Anyway, this means we're back on 1.4.2 as a hotfix, meaning no intents.
However we still get the privileged intents because we checked them in
the dev portal and the announcement clarifies if you don't specify
intents (meaning, using discord.py 1.4.2) you'll get the events based
on what you checked in the dev portal.