Make starboarding_messages a list instead of a set
The *actual* cause of the issues I mentioned in 379318b3 was that
Message objects could simply not be checked for uniqueness anymore,
like I assumed we could by having a set of messages. That's the only
thing that was broken in 1.5.0 by not making messages hashable
anymore - I confused object comparison in Python with Java (and maybe
some other languages) where - as far as I can recall at the moment -
hashes are always used for comparing two objects. However, in Python
this change would only break trying to index objects, or enforcing
uniqueness in a set.