Use discord.Client.cached_messages wherever possible
Looks like discord.py has added a cached_messages attribute to the
discord.Client object, letting you read the cached messages your client
has. But it only lets you read it, you can't write to it - so we'll
still have to use discord.Client._connection._messages for that one.
However, we'll use cached_messages as it is the more intended and
officially supported way to read cached messages.