Make errors in console clearer, add \errors
Unhandled exceptions in events now have a red "banner" above them (with
ANSI escape codes just like the "fault did occur in command" logs
already got earlier) and now have the time the error occurred, instead
of just the default "Ignoring exception in ..." message from
discord.py.

Unhandled exceptions are now also added to a list, which can be read by
the \errors command. This gives a basic representation of the last
exceptions that occurred (the time, the event in which it was raised
and the exception type name). It also gives how many exceptions were
logged in the last 24 hours in total.

The amount of information given by \errors may not be enough for
debugging issues in detail, but it's going to be incredibly useful to
see if anything's going wrong at a glance without having to scroll back
in the tmux buffer and use the timestamps from "Syncing invite cache"
messages as your only reference. I also remember we once decided to
make the bot never automatically display detailed error messages
because there may be a risk of accidentally leaking secrets like the
bot token in unforseen ways, but this command only outputs method and
type names.