Don't require Ban Members permission to log a ban
The bot would fail with an unhandled exception if it didn't have the Ban
Members permission at the time someone got banned and it attempted to
log the ban. This is because the bot uses the ban list itself to grab
the reason for the ban, but if the bot doesn't have the Ban Members
permission, then this will fail, and the log won't be posted.

To fix this, the bot will just ignore it if it can't grab the ban list
for whatever reason. It'll say "(cannot be determined)", which doesn't
indicate that it requires a Ban Members permission, but at least it's
better to post a message than to not.

Also, as a backup, the bot COULD grab the reason from the audit logs,
but that can be done in a later commit.